return globs instead of all existing files
This reduces the amount of strings (files) we send over the wire, and it
also allows us to watch new files in a given folder with the known
extensions.

Adding a new file in a new folder or adding a new file in a folder with
a different extension as the other files won't work and a restart of the
server will be required.